fixed some s3 bugs

+++ controllers/admin/controllerPhrases.js
@@ -80,19 +80,20 @@ exports.update = (req, res) => {
             return res.end('Error uploading file.')
         } else console.log('Image updated!')
 
+        console.log(req.body.deleteImage)
+
         phrasesDB.getImageUrl(req.params.id)
             .then(obj => {
-                if (obj && obj.url !== '') {
+                if (obj && obj.url !== '' && req.body.deleteImage == 1) {
                     // getImageNameFromUrl
                     const tmp = obj.url.split('/')
                     s3.deleteImage(tmp[tmp.length - 1])
                 }
                 const date = new Date().toISOString().split('T')[0] // 2021-05-26T21:53:36.244Z
-                console.log(date)
                 const phrase = {
                         id: req.params.id,
                         text: req.body.text,
-                        img: (req.file) ? req.file.location : '',
+                        img: (req.body.deleteImage == 0) ? req.body.oldImgUrl : (req.file) ? req.file.location : '',
                         quotedById: req.body.quotedById,
                         authorId: req.user.id,
                         isFinished: (req.body.isFinished === 'on') ? 1 : 0,
